+++ NeoGebra/src/UI/ButtonUI.cpp
@@ -4,16 +4,16 @@
 #include "Rendering/Renderer.h"
 #include "Application.h"
 
-ButtonUI::ButtonUI(double leftX, double rightX, double topY, double bottomY, void(*func)(void*), void* obj, const std::string& text)
+ButtonUI::ButtonUI(float leftX, float rightX, float topY, float bottomY, void(*func)(void*), void* obj, const std::string& text)
 	: ButtonUI(leftX, rightX, topY, bottomY, func, obj, AdvancedString(text)) {}
 
-ButtonUI::ButtonUI(double leftX, double rightX, double topY, double bottomY, void(*func)(void*), void* obj, const AdvancedString& text)
+ButtonUI::ButtonUI(float leftX, float rightX, float topY, float bottomY, void(*func)(void*), void* obj, const AdvancedString& text)
 	: UIElement(leftX, rightX, topY, bottomY, "ButtonUI"),
 	m_Background(std::make_shared<Square>(leftX, rightX, topY, bottomY, std::array{ 0.0f, 0.6f, 1.0f, 1.0f })),
 	m_Obj(obj)
 {
 	m_Action = func;
-	m_Texts.push_back(std::make_shared<Text>(text, leftX + 0.005f, rightX, bottomY + 0.04f, 36));
+	m_Texts.push_back(std::make_shared<Text>(text, leftX + 0.005f, rightX, bottomY + 0.04f, 36.0f));
 }
 
 ButtonUI::~ButtonUI()
