+++ NeoGebra/src/UI/ButtonUI.h
@@ -9,8 +9,8 @@
 class ButtonUI : public UIElement
 {
 public:
-	ButtonUI(double leftX, double rightX, double topY, double bottomY, void(*func)(void*), void* obj, const std::string& text);
-	ButtonUI(double leftX, double rightX, double topY, double bottomY, void(*func)(void*), void* obj, const AdvancedString& text);
+	ButtonUI(float leftX, float rightX, float topY, float bottomY, void(*func)(void*), void* obj, const std::string& text);
+	ButtonUI(float leftX, float rightX, float topY, float bottomY, void(*func)(void*), void* obj, const AdvancedString& text);
 	~ButtonUI();
 
 	virtual void RenderPass(Renderer* r) override;
